+++ NEWS.md
@@ -6,6 +6,7 @@
 - added `power_method()`, power method for dominant eigenvector (thx: Gaston Sanchez)
 - added `arrows3d()` for 3D geometric diagrams
 - added `vectors3d()` for 3D geometric diagrams
+- added `corner()` for 2D, 3D geometric diagrams
 
 
 # matlib 0.5.2